IGNITE-20978 / turn executor collection to executor
